requirements.txt
Django==4.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt
social-auth-app-django
social-auth-core
psycopg2-binary
python-dotenv
drf-yasg
django-currentuser
django-simple-history



#configurar la seguridad
